fix: grid-stack implementation
