+++ src/app/view/base/grid-stack/grid-stack.tsx
@@ -1,69 +1,29 @@
-import { GridStack as GS, GridStackOptions, GridStackWidget } from 'gridstack';
+import { GridStack as GS } from 'gridstack';
 import React from 'react';
-import { useMount, usePrevious, useUnmount } from 'react-use';
-import * as R from 'remeda';
-import { useGridStackContext } from '.';
 import { Widget } from './_type';
 import { GridStackItem } from './grid-stack-item';
+import { useGridStackContext } from './gs-context';
 
 import 'gridstack/dist/gridstack.min.css';
 
 type DivProps = React.ComponentProps<'div'>;
-
-export type OnChangeFn = (grid: GS, items: GridStackWidget[]) => void;
-export type OnSizeChangeFn = (grid: GS, cols: number) => void;
-export type RenderItemFn = (widget: Widget) => DivProps;
+export type RenderItemFn = (grid: GS, widget: Widget) => DivProps;
 
 export interface GridStackProps {
   widgets: Widget[];
-  options?: GridStackOptions;
-  /** Fired when any widget is moved or resized. */
-  onChange?: OnChangeFn;
-  /** Fired when grid container size changes */
-  onSizeChange?: OnSizeChangeFn;
   /** Passed to each widget to render additional components in each grid item */
   renderItem?: RenderItemFn;
 }
 
 export const GridStack: React.FC<GridStackProps> = ({
   widgets,
-  options,
-  onChange,
-  onSizeChange,
   renderItem,
 }) => {
-  const { containerNode, gsRef, grid, setGrid } = useGridStackContext();
+  const { grid } = useGridStackContext();
   const itemElsRef = React.useRef(new Map<string, HTMLDivElement>());
   const prevIdsRef = React.useRef<string[]>([]);
-  const prevOptions = usePrevious(options);
-
-  useMount(() => {
-    if (!containerNode) {
-      return;
-    }
-
-    const gs = intializeGridStack(containerNode, {
-      options,
-      onChange,
-    });
-    gsRef.current = gs;
-    setGrid(gs);
-  });
-
-  useUnmount(() => {
-    const gs = gsRef.current;
-    if (gs) {
-      gs.destroy(false);
-      gsRef.current = null;
-    }
-  });
 
   React.useEffect(() => {
-    const gs = gsRef.current;
-    if (!gs) {
-      return;
-    }
-
     const currentIds = new Set(widgets.map((w) => w.id));
 
     // Remove widgets that no longer needs to be rendered
@@ -72,12 +32,12 @@ export const GridStack: React.FC<GridStackProps> = ({
       .forEach((id) => {
         const el = itemElsRef.current.get(id);
         if (el) {
-          gs.removeWidget(el, false);
+          grid.removeWidget(el, false);
           itemElsRef.current.delete(id);
         }
       });
 
-    gs.batchUpdate(true);
+    grid.batchUpdate(true);
     widgets.forEach(({ content: _content, ...widgetOpts }) => {
       const { id } = widgetOpts;
       const el = itemElsRef.current.get(id);
@@ -86,15 +46,15 @@ export const GridStack: React.FC<GridStackProps> = ({
       }
       const isNew = !prevIdsRef.current.includes(id);
       if (isNew) {
-        gs.makeWidget(el, widgetOpts);
+        grid.makeWidget(el, widgetOpts);
       } else {
-        gs.update(el, widgetOpts);
+        grid.update(el, widgetOpts);
       }
     });
-    gs.batchUpdate(false);
+    grid.batchUpdate(false);
 
     prevIdsRef.current = widgets.map((w) => w.id);
-  }, [containerNode, gsRef, widgets]);
+  }, [grid, widgets]);
 
   const initItemRefs = React.useCallback(
     (el: HTMLDivElement | null, id: string) => {
@@ -105,63 +65,14 @@ export const GridStack: React.FC<GridStackProps> = ({
     []
   );
 
-  // Reinitialize grid instance when options are modified
-  React.useLayoutEffect(() => {
-    let gs = gsRef.current;
-    if (!gs || !containerNode || R.isDeepEqual(prevOptions, options)) {
-      return;
-    }
-
-    gs.removeAll(false);
-    gs.destroy(false);
-    gs = intializeGridStack(containerNode, {
-      options,
-      onChange,
-    });
-    gsRef.current = gs;
-    setGrid(gs);
-  }, [containerNode, gsRef, onChange, prevOptions, options, setGrid]);
-
-  React.useEffect(() => {
-    const gs = gsRef.current;
-    if (!gs || !containerNode) {
-      return;
-    }
-
-    const update = () => {
-      onSizeChange?.(gs, gs.getColumn());
-    };
-    const observer = new ResizeObserver(update);
-    observer.observe(containerNode);
-    update();
-
-    return () => observer.disconnect();
-  }, [containerNode, gsRef, onSizeChange]);
-
   return widgets.map((widget) => {
     return (
       <GridStackItem
         key={widget.id}
         ref={(el) => initItemRefs(el, widget.id)}
         widget={widget}
-        {...renderItem?.(widget)}
+        {...renderItem?.(grid, widget)}
       />
     );
   });
 };
-
-interface InitGridStackOpt {
-  options?: GridStackOptions;
-  onChange?: OnChangeFn;
-}
-function intializeGridStack(
-  containerNode: HTMLDivElement,
-  opt?: InitGridStackOpt
-) {
-  const gs = GS.init(opt?.options, containerNode);
-
-  gs.on('change', (_event, items) => {
-    opt?.onChange?.(gs, items as GridStackWidget[]);
-  });
-  return gs;
-}
