+++ src/app/view/base/grid-stack/index.tsx
@@ -1,5 +1,6 @@
 export { type GridStackOptions, type GridStackWidget } from 'gridstack';
 export * from './_type';
 export * from './grid-stack';
+export * from './gs-container';
 export * from './gs-context';
 export * from './widget-util';
